CARDS-1475: PROMs dashboard Put all the frontend sources in a /proms/ subdir Rename the Dashboard to PromsDashboard
