+++ proms-resources/frontend/src/main/frontend/src/proms/index.jsx
@@ -28,7 +28,7 @@ import {
 } from "@material-ui/core";
 import { createBrowserHistory } from "history";
 import { MuiThemeProvider } from '@material-ui/core/styles';
-import { appTheme } from "./themePalette.jsx";
+import { appTheme } from "../themePalette.jsx";
 import QuestionnaireSet from "./QuestionnaireSet.jsx";
 import PatientIdentification from "./PatientIdentification.jsx";
 import PromsFooter from "./Footer.jsx";
