+++ proms-resources/frontend/src/main/frontend/webpack.config.js
@@ -6,11 +6,11 @@ module_name = require("./package.json").name + ".";
 module.exports = {
   mode: 'development',
   entry: {
-    [module_name + 'index']: './src/index.jsx',
-    [module_name + 'ToULink']: './src/ToULink.jsx'
-    [module_name + 'Dashboard']: './src/Dashboard.jsx',
-    [module_name + 'PromsView']: './src/PromsView.jsx',
-    [module_name + 'VisitView']: './src/VisitView.jsx',
+    [module_name + 'index']: './src/proms/index.jsx',
+    [module_name + 'ToULink']: './src/proms/ToULink.jsx'
+    [module_name + 'PromsDashboard']: './src/proms/PromsDashboard.jsx',
+    [module_name + 'PromsView']: './src/proms/PromsView.jsx',
+    [module_name + 'VisitView']: './src/proms/VisitView.jsx',
     [module_name + 'pmccIcon']: '@material-ui/icons/Favorite.js'
   },
   plugins: [
